Move tessellation::PolylineEvents into path_iterator::FromPolyline.
